requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.15.2
idna==2.8
isort==4.3.21
jmespath==0.9.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.18.1
pathspec==0.7.0
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
regex==2020.1.8
requests==2.22.0
s3transfer==0.3.2
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
